refactor: 은행 선택 컴포넌트 파일 분리
