+++ app/src/main/java/woowacourse/payments/ui/addcard/BankSelectBottomSheet.kt
@@ -1,0 +1,23 @@
+package woowacourse.payments.ui.addcard
+
+import androidx.compose.material3.ExperimentalMaterial3Api
+import androidx.compose.material3.ModalBottomSheet
+import androidx.compose.material3.rememberModalBottomSheetState
+import androidx.compose.runtime.Composable
+import woowacourse.payments.domain.BankType
+
+@OptIn(ExperimentalMaterial3Api::class)
+@Composable
+fun BankSelectBottomSheet(
+    onDismiss: () -> Unit,
+    onBankSelected: (BankType) -> Unit,
+) {
+    val modalBottomSheetState = rememberModalBottomSheetState(skipPartiallyExpanded = true)
+
+    ModalBottomSheet(
+        sheetState = modalBottomSheetState,
+        onDismissRequest = onDismiss,
+    ) {
+        BankSelectRow(onClick = onBankSelected)
+    }
+}
