+++ app/src/main/java/woowacourse/payments/ui/addcard/BankSelectBtn.kt
@@ -1,0 +1,42 @@
+package woowacourse.payments.ui.addcard
+
+import androidx.compose.foundation.Image
+import androidx.compose.foundation.clickable
+import androidx.compose.foundation.layout.Arrangement
+import androidx.compose.foundation.layout.Column
+import androidx.compose.foundation.layout.height
+import androidx.compose.foundation.layout.size
+import androidx.compose.foundation.layout.width
+import androidx.compose.material3.Text
+import androidx.compose.runtime.Composable
+import androidx.compose.ui.Alignment
+import androidx.compose.ui.Modifier
+import androidx.compose.ui.res.painterResource
+import androidx.compose.ui.unit.dp
+import androidx.compose.ui.unit.em
+import androidx.compose.ui.unit.sp
+import woowacourse.payments.domain.BankType
+
+@Composable
+fun BankSelectBtn(
+    bank: BankType,
+    onClick: () -> Unit,
+    modifier: Modifier = Modifier,
+) {
+    Column(
+        modifier =
+            modifier
+                .width(69.dp)
+                .height(65.dp)
+                .clickable(onClick = onClick),
+        horizontalAlignment = Alignment.CenterHorizontally,
+        verticalArrangement = Arrangement.Center,
+    ) {
+        Image(
+            modifier = Modifier.size(37.dp),
+            painter = painterResource(bank.bankLogo),
+            contentDescription = "${bank.bankName} 로고",
+        )
+        Text(bank.bankName, fontSize = 16.sp, letterSpacing = (-0.085).em)
+    }
+}
